bring this shit with writing in db to work; But we can't take any skill from skills list, but we can it from a charlist direct

+++ src/main/java/org/waaagh/repositories/CharListRepository.java
@@ -1,17 +1,14 @@
 package org.waaagh.repositories;
 
-import org.springframework.data.jpa.repository.Query;
-import org.waaagh.model.CharList;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import org.waaagh.model.CharList;
 
 import java.util.List;
 import java.util.Optional;
 
 @Repository
 public interface CharListRepository extends JpaRepository<CharList, Integer> {
-//    @Query("SELECT v FROM Visit v WHERE v.visitType='NEW'")
     List<CharList> findAll();
-
-    Optional<CharList> findById(Integer id);
+    Optional<CharList> findById(Long id);
 }
