+++ src/main/java/org/waaagh/service/CharListService.java
@@ -5,9 +5,9 @@ import org.waaagh.model.CharList;
 public interface CharListService {
     void addCharList(CharList charList);
 
-    void removeCharList(Integer id);
+    void removeCharList(Long id);
 
     void updateCharList(CharList charList);
 
-    CharList getCharListById(Integer id);
+    CharList getCharListById(Long id);
 }
